Change Block::version to 16 bits and mine genesis

+++ src/params.h
@@ -30,7 +30,7 @@ public:
     };
 
     // consensus parameter setting
-    uint32_t version;
+    uint16_t version;
     uint32_t targetTimespan;
     uint32_t timeInterval;
     uint32_t interval;
